+++ TheSocial_Post/Controllers/PostController.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using System.IdentityModel.Tokens.Jwt;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -16,6 +13,7 @@ namespace TheSocial_Post.Controllers
     public class PostController : ControllerBase
     {
         private readonly IPostService _postService;
+
         private readonly IMapper _mapper;
         private readonly ResponseDto _response;
 
@@ -26,10 +24,19 @@ namespace TheSocial_Post.Controllers
             _mapper = mapper;
             _response = new ResponseDto();
         }
+        private string GetUserIdFromToken()
+        {
+            var token = Request.Headers["Authorization"].ToString();
+            var decodedToken = new JwtSecurityTokenHandler().ReadJwtToken(token.Split(" ")[1]);
+            return decodedToken.Claims.FirstOrDefault(c => c.Type == "sub")?.Value;
+        }
+
         [HttpPost]
         public async Task<ActionResult<ResponseDto>> CreatePost(PostRequestDto postRequestDto)
         {
             var newPost = _mapper.Map<Post>(postRequestDto);
+            var userId = GetUserIdFromToken();
+            newPost.UserId = Guid.Parse(userId);
             var response = await _postService.CreatePostAsync(newPost);
             if (response != null)
             {
@@ -56,8 +63,9 @@ namespace TheSocial_Post.Controllers
             _response.Message = "Something went wrong";
             return BadRequest(_response);
         }
-        // get post by id and also include comments
+        // get post by id
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<ResponseDto>> GetPostById(Guid id)
         {
             var post = await _postService.GetPostByIdAsync(id);
@@ -72,36 +80,21 @@ namespace TheSocial_Post.Controllers
             _response.Message = "Something went wrong";
             return BadRequest(_response);
         }
-        // <!-- delete post by id --> did not work
-        // public async Task<ActionResult<ResponseDto>> DeletePost(Guid id)
-        // {
-        //     var ownerId = User.Claims.FirstOrDefault(c => c.Type == "Id");
-        //     var post = await _postService.GetPostByIdAsync(id);
-        //     // Check if the post exists and the UserId in the token matches the post's UserId
-        //     if (post != null && ownerId.Value == post.UserId.ToString())
-        //     {
-        //         var response = await _postService.DeletePostAsync(post);
-        //         if (response != null)
-        //         {
-        //             _response.IsSuccess = true;
-        //             _response.Message = "Successfully deleted";
-        //             return Ok(_response);
-        //         }
-        //         _response.IsSuccess = false;
-        //         _response.Message = "Something went wrong";
-        //         return BadRequest(_response);
-        //     }
-        //     _response.IsSuccess = false;
-        //     _response.Message = "Post not found";
-        //     return BadRequest(_response);
-        // }
+        // delete post and authorize the owner of the post
         [HttpDelete("{id}")]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public async Task<ActionResult<ResponseDto>> DeletePost(Guid id)
         {
             var post = await _postService.GetPostByIdAsync(id);
+            var userId = GetUserIdFromToken();
             if (post != null)
             {
+                if (post.UserId.ToString() != userId)
+                {
+                    _response.IsSuccess = false;
+                    _response.Message = "You are not authorized to delete this post";
+                    return BadRequest(_response);
+                }
                 var response = await _postService.DeletePostAsync(post);
                 if (response != null)
                 {
@@ -124,7 +117,7 @@ namespace TheSocial_Post.Controllers
         public async Task<ActionResult<ResponseDto>> GetAllPostsByUserId(Guid userId)
         {
             var posts = await _postService.GetPostsByUserIdAsync(userId);
-            if (posts != null)
+            if (posts.Count() > 0)
             {
                 _response.Message = "";
                 _response.IsSuccess = true;
@@ -132,30 +125,40 @@ namespace TheSocial_Post.Controllers
                 return Ok(_response);
             }
             _response.IsSuccess = false;
-            _response.Message = "Something went wrong";
+            _response.Message = "The user does not have any posts yet or does not exist!!";
             return BadRequest(_response);
         }
         [HttpPut]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public async Task<ActionResult<ResponseDto>> UpdatePost(Guid id, PostRequestDto postRequestDto)
         {
             var post = await _postService.GetPostByIdAsync(id);
-            if (post == null)
+            var userId = GetUserIdFromToken();
+            //check if the post exists and the UserId in the token matches the post's UserId
+            if (post != null)
             {
+                if (post.UserId.ToString() != userId)
+                {
+                    _response.IsSuccess = false;
+                    _response.Message = "You are not authorized to update this post";
+                    return BadRequest(_response);
+                }
+                //update post
+                var postToUpdate = _mapper.Map(postRequestDto, post);
+                var response = await _postService.UpdatePostAsync(postToUpdate);
+                if (response != null)
+                {
+                    _response.IsSuccess = true;
+                    _response.Message = "Successfully updated";
+                    return Ok(_response);
+                }
                 _response.IsSuccess = false;
-                _response.Message = "Post not found";
+                _response.Message = "Something went wrong";
                 return BadRequest(_response);
             }
-            var postToUpdate = _mapper.Map(postRequestDto, post);
-            var response = await _postService.UpdatePostAsync(postToUpdate);
-            if (response != null)
-            {
-                _response.IsSuccess = true;
-                _response.Message = "Successfully updated";
-                return Ok(_response);
-            }
+            //if not authorized return an error
             _response.IsSuccess = false;
-            _response.Message = "Something went wrong";
+            _response.Message = "Post not found";
             return BadRequest(_response);
         }
